- more warning fixes in 'core'.

+++ source/core/console/c_notifybuffer.cpp
@@ -116,10 +116,10 @@ void FNotifyBuffer::DrawNative()
 
 	FFont* font = isBlood() ? SmallFont2 : SmallFont;
 
-	int line = isBlood() ? Top : (g_gameType & GAMEFLAG_SW) ? 40 : font->GetDisplacement();
+	double line = isBlood() ? Top : (g_gameType & GAMEFLAG_SW) ? 40 : font->GetDisplacement();
 	bool canskip = isBlood();
 	double scale = 1 / (NotifyFontScale * con_notifyscale);
-	int lineadv = font->GetHeight() / NotifyFontScale;
+	double lineadv = font->GetHeight() / NotifyFontScale;
 
 	for (unsigned i = topline; i < Text.Size(); ++i)
 	{
